+++ lwb/metalang/cfg/cfg/src/main/java/mb/cfg/metalang/CfgStrategoConfig.java
@@ -4,7 +4,6 @@ import mb.cfg.CompileMetaLanguageSourcesShared;
 import mb.resource.hierarchical.ResourcePath;
 import mb.spoofax.compiler.adapter.ExportsCompiler;
 import mb.spoofax.compiler.language.StrategoRuntimeLanguageCompiler;
-import mb.spoofax.compiler.util.BuilderBase;
 import mb.spoofax.compiler.util.Conversion;
 import mb.spoofax.compiler.util.Shared;
 import org.immutables.value.Value;
@@ -20,50 +19,31 @@ public interface CfgStrategoConfig extends Serializable {
     String exportsId = "Stratego";
 
 
-    class Builder extends ImmutableCfgStrategoConfig.Builder implements BuilderBase {
-        static final String propertiesPrefix = "stratego.";
-        static final String languageStrategyAffix = propertiesPrefix + "languageStrategyAffix";
-
-        public Builder withPersistentProperties(Properties properties) {
-            with(properties, languageStrategyAffix, this::languageStrategyAffix);
-            return this;
-        }
-    }
+    class Builder extends ImmutableCfgStrategoConfig.Builder {}
 
     static Builder builder() {return new Builder();}
 
 
     @Value.Default default CfgStrategoSource source() {
         return CfgStrategoSource.files(CfgStrategoSource.Files.builder()
-            .compileMetaLanguageSourcesShared(compileLanguageShared())
+            .compileMetaLanguageSourcesShared(compileMetaLanguageSourcesShared())
+            .shared(shared())
             .build()
         );
     }
 
 
-    @Value.Default default boolean enableSdf3StatixExplicationGen() {
-        // TODO: move into source after CC lab.
-        return false;
-    }
-
-    @Value.Default default String languageStrategyAffix() {
-        // TODO: should go into CfgStrategoSource.Files? complicated due to persistent properties though...
-        // TODO: convert to Stratego ID instead of Java ID.
-        return Conversion.nameToJavaId(shared().name().toLowerCase());
-    }
-
-
     @Value.Default default ResourcePath javaSourceFileOutputDirectory() {
         // Generated Java sources directory, so that Gradle compiles the Java sources into classes.
-        return compileLanguageShared().generatedJavaSourcesDirectory();
+        return compileMetaLanguageSourcesShared().generatedJavaSourcesDirectory();
     }
 
     @Value.Default default ResourcePath javaClassFileOutputDirectory() {
-        return compileLanguageShared().languageProject().project().buildClassesDirectory();
+        return compileMetaLanguageSourcesShared().languageProject().project().buildClassesDirectory();
     }
 
     @Value.Default default String outputJavaPackageId() {
-        return compileLanguageShared().languageProject().packageId() + ".strategies";
+        return compileMetaLanguageSourcesShared().languageProject().packageId() + ".strategies";
     }
 
     default String outputJavaPackagePath() {
@@ -71,7 +51,7 @@ public interface CfgStrategoConfig extends Serializable {
     }
 
     default String outputLibraryName() {
-        return compileLanguageShared().languageProject().project().coordinate().artifactId;
+        return compileMetaLanguageSourcesShared().languageProject().project().coordinate().artifactId;
     }
 
     default ResourcePath outputJavaInteropRegistererFile() {
@@ -85,13 +65,13 @@ public interface CfgStrategoConfig extends Serializable {
 
     /// Automatically provided sub-inputs
 
-    CompileMetaLanguageSourcesShared compileLanguageShared();
+    CompileMetaLanguageSourcesShared compileMetaLanguageSourcesShared();
 
     Shared shared();
 
 
     default void savePersistentProperties(Properties properties) {
-        properties.setProperty(Builder.languageStrategyAffix, languageStrategyAffix());
+        source().getFiles().savePersistentProperties(properties);
     }
 
 
@@ -101,6 +81,6 @@ public interface CfgStrategoConfig extends Serializable {
     }
 
     default void syncTo(ExportsCompiler.Input.Builder builder) {
-        source().getFiles().exportDirectories().forEach(exportDirectory -> builder.addDirectoryExport(exportsId, exportDirectory));
+        source().getFiles().syncTo(builder);
     }
 }